--- FastSimulation/CTPPSRecHitProducer/BuildFile.xml
@@ -1,11 +1,0 @@
-<use name="FWCore/Framework"/>
-<use name="FWCore/PluginManager"/>
-<use name="FWCore/ParameterSet"/>
-<use name="FWCore/ServiceRegistry"/>
-<use name="SimDataFormats/TrackingHit"/>
-<use name="SimDataFormats/CrossingFrame"/>
-<use name="DataFormats/Common"/>
-<use name="clhep"/>
-<use name="SimDataFormats/GeneratorProducts"/>
-<use name="hepmc"/>
-<use name="FastSimulation/CTPPSFastGeometry"/>
